python/pymortar/pyproject.toml
[tool.poetry]
name = "pymortar"
version = "0.2.2"
description = ""
authors = ["Gabe Fierro <gtfierro@cs.berkeley.edu>"]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.18"
googleapis-common-protos = "^1.5"
grpcio-tools = "^1.18"
pandas = "^0.23.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
